[TEST] Scratch of the test to deploy version 2.0
Correct ARGUMENT.get to return the correct type (default is to return
a string). Join the experiment tests into correspondent model test
file. Tests related to gaussian model are stored in
'test_gaussianModelBuild.py' and tests related to the multi layer
model are stored in 'test_multiLayerModelBuild.py'.
