+++ TDD/test_gaussianModelBuild.py
@@ -24,7 +24,7 @@ import os
 # Change to modeling directory
 currentDir = os.getcwd()
 path = os.path.join(currentDir,os.pardir)
-path = os.path.abspath(path)+'/experiments/modelingAndPEFInterpolation/'
+path = os.path.abspath(path)+'/experiments/'
 os.chdir(path)
 
 class TestScons(unittest.TestCase):
@@ -37,7 +37,7 @@ class TestScons(unittest.TestCase):
 		Test the building of the gaussian reflector in
 		a linear velocity model
 		'''
-		result = subprocess.call('cd gaussianModel && scons dataCube.rsf',shell=True)
+		result = subprocess.call('cd modelingAndPEFInterpolation/gaussianModel && scons dataCube.rsf',shell=True)
 		self.assertEqual(result,0,'FAILED: gaussian reflector modeling failed!')
 
 
@@ -46,5 +46,22 @@ class TestScons(unittest.TestCase):
 		Test of the PEF interpolation step of the seismic data cube
 		modeled with the gaussian reflector linear velocity model.
 		'''
-		result = subprocess.call('cd gaussianModel && scons -j3 nhi=1',shell=True)
+		result = subprocess.call('cd modelingAndPEFInterpolation/gaussianModel && scons -j3 nhi=1',shell=True)
 		self.assertEqual(result,0,'FAILED: gaussian reflector PEF interpolation failed!')
+
+	def test_fullInterpolationAndStack(self):
+		'''
+		Unit test of the fullInterpolationAndStack experiment
+		'''
+
+		result = subprocess.call('cd fullInterpolationAndStack && scons -j3 nm0=2 nt0=3',shell=True)
+		self.assertEqual(result,0,'FAILED: experiment fullInterpolationAndStack failed!')
+
+	def test_fullInterpolationAndStackCDS(self):
+		'''
+		Unit test of the fullInterpolationAndStackCDS experiment
+		'''
+		result = subprocess.call('cd fullInterpolationAndStackCDS && scons -j3 nm0=2 nt0=3',shell=True)
+		self.assertEqual(result,0,'FAILED: experiment fullInterpolationAndStackCDS failed!')
+
+
