+++ experiments/modelingAndPEFInterpolation/multiLayerModel/SConstruct
@@ -61,10 +61,10 @@ kinewmod(reflectors='interfaces',
 
 # Use default value or pass through command line
 # PEF coeficients and PEF Smooth radius
-a1=ARGUMENTS.get('a1',10)
-a2=ARGUMENTS.get('a2',2)
-rect1=ARGUMENTS.get('rect1',50)
-rect2=ARGUMENTS.get('rect2',2)
+a1=int(ARGUMENTS.get('a1',10))
+a2=int(ARGUMENTS.get('a2',2))
+rect1=int(ARGUMENTS.get('rect1',50))
+rect2=int(ARGUMENTS.get('rect2',2))
 
 # Number of offsets to interpolate
 nhi=ARGUMENTS.get('nhi',1)
